+++ hailstorm-web-client/src/AWSForm/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "AWSForm.tsx"
+}
